==> Proof-of-Concept/2. Saving Stay_duration_dict.ipynb <==
import glob
import pickle
import numpy as np
import pandas as pd

all_files=glob.glob("./Trails/*/*/mu_serv_*.csv")

test_dates=['2019-06-28','2019-06-30','2019-07-01','2019-07-02','2019-07-03','2019-07-04','2019-07-05']

#train files
train_files=[f for f in all_files if f.split('_')[-1].split(".")[0] not in test_dates]

#test files
test_files=[f for f in all_files if f.split('_')[-1].split(".")[0] in test_dates]

# Training Data

df=pd.concat([pd.read_csv(f) for f in train_files])

dic={}
for poi_column in ['Is_Bus_stop', 'Is_Turn', 'Is_Signal', 'Is_Congestion', 'Is_Adhoc']:
    dic[poi_column]=dict(df[df[poi_column]==1].groupby('time_zone')['stay_duration'].mean())

#Saving stay_duration_dict
with open('./logs/stay_duration.pickle', 'wb') as handle:
    pickle.dump(dic, handle, protocol=pickle.HIGHEST_PROTOCOL)

#NICE
