==> Proof-of-Concept/3. Writing Estimated Stay time to mu_serv files.ipynb <==
import glob
import pickle
import numpy as np
import pandas as pd

all_files=glob.glob("./Trails/*/*/mu_serv_*.csv")

#loading stay_duration_dict
with open('./logs/stay_duration.pickle', 'rb') as handle:
    dic= pickle.load(handle)

dic

# Writing Estimated stay Time

def get_estimated_stay_duration(time_zone,is_bus,is_turn,is_signal,is_congestion,is_adhoc):
    l=[]
    if is_bus==1:
        l.append(dic['Is_Bus_stop'][time_zone])
    if is_turn==1:
        l.append(dic['Is_Turn'][time_zone])
    if is_signal==1:
        l.append(dic['Is_Signal'][time_zone])
    if is_congestion==1:
        l.append(dic['Is_Congestion'][time_zone])
    if is_adhoc==1 or len(l)==0:
        l.append(dic['Is_Adhoc'][time_zone])
    return max(l)

for f in all_files:
    df=pd.read_csv(f)
    df['estimated_stay_duration']=df[['time_zone']+list(dic.keys())].apply(lambda tup:get_estimated_stay_duration(*tup),axis=1)
    df.to_csv(f,index=False)

#NICE
