==> Proof-of-Concept/4. Predict Next Stop Time From Previous point.ipynb <==
import glob
import datetime
import numpy as np
import pandas as pd
from library.preprocessing import distance
from geopy.distance import geodesic

def geodistance(pointA, pointB):
    return geodesic(pointA, pointB).meters

all_folders=glob.glob("./Trails/*/*/")

test_dates=['2019-06-28','2019-06-30','2019-07-01','2019-07-02','2019-07-03','2019-07-04','2019-07-05']
formatted_test_dates=['-'.join(d.split("-")[::-1]) for d in test_dates] #reversing order

get_date=lambda fname:fname.split("\\")[-2].split("_")[0]
get_format_date=lambda fname: '-'.join(get_date(fname).split("-")[::-1])

#test folders
test_file_pairs=\
[(f+"GPS.csv",
 f+"mu_serv_{}.csv".format(get_format_date(f))) for f in all_folders if get_date(f) in formatted_test_dates]

def get_Sec_from_datetime(dt):
    timeStamp=pd.to_datetime(dt,format="%m/%d/%Y %H:%M:%S")
    return timeStamp.timestamp()

def process_GPS(f_name):
    df=pd.read_csv(f_name)
    start_lat,start_long,start_time=df[['#lat','long','time']].iloc[0]

    next_hop_distance=[]
    time_elapsed=[]
    for next_lat,next_long,next_time in df[['#lat','long','time']].values:
        next_hop_distance.append(geodistance((start_lat,start_long),(next_lat,next_long)))
        time_elapsed.append(get_Sec_from_datetime(next_time)-get_Sec_from_datetime(start_time)+1e-9)
        start_lat,start_long,start_time=next_lat,next_long,next_time

    #next_hop_distance
    #time_elapsed
    df['next_hop_distance']=next_hop_distance
    df['time_elapsed']=time_elapsed
    df['start_time']=df.time.apply(lambda e:e.split(" ")[1])
    return df[['start_time','#lat','long','next_hop_distance','time_elapsed']].copy()

train_initial_speed=[]
tzones=[]
train_files=\
[(f+"GPS.csv",
 f+"mu_serv_{}.csv".format(get_format_date(f))) for f in all_folders if get_date(f) not in formatted_test_dates]

for gps_f,mu_f in train_files:
    tz=pd.read_csv(mu_f).time_zone[0]
    df=process_GPS(gps_f).iloc[:5*60]#5 min data
    tzones.append(tz)
    train_initial_speed.append(df.next_hop_distance.sum()/df.time_elapsed.sum())

init_speed_dict=\
pd.DataFrame({'time_zone':tzones,'speed':train_initial_speed}).groupby("time_zone")['speed'].mean()

dict(init_speed_dict)

#{'Afternoon': 3.208857802457898,
#'Early_Morning': 2.4823871880135058,
#'Evening': 1.9810927189917138,
#'Morning': 3.0719420700235105}

def process_MU(f_name):
    df_mu=pd.read_csv(f_name)
    init_speed_dict={'Afternoon': 3.208857802457898,
                     'Early_Morning': 2.4823871880135058,
                     'Evening': 1.9810927189917138,
                     'Morning': 3.0719420700235105}
    initial_speed=init_speed_dict[df_mu.time_zone[0]] #calcullated from past data
    data=df_mu[['start_time','lat','long','estimated_stay_duration','stay_duration']]
    label=df_mu[['Is_Bus_stop', 'Is_Turn', 'Is_Signal', 'Is_Congestion', 'Is_Adhoc','time_zone']]
    return data,label,initial_speed

def get_time_plus_delta(TIME,DELTA):
    return str((pd.to_datetime(TIME)+datetime.timedelta(seconds=DELTA)).time())

def get_immediate_predicted_arrival_time(file):
    cal_speed_on_past_min=20 ##speed is calculated from past 20 min trail
    
    df_gps=process_GPS(file[0])
    df_mu,mu_labels,speed=process_MU(file[1])

    start_time,start_lat,start_long,start_estimated_stay_duration,start_stay_duration=df_mu.iloc[0]

    estimated_start_time=['NA']
    error_in_mins=[np.nan]
    SPEED=[speed]
    total_stay=0

    for next_start_time,next_lat,next_long,next_estimated_stay_duration,next_stay_duration in df_mu.iloc[1:].values:

        #immediate patch distance
        patch_distance=df_gps[(df_gps.start_time<=next_start_time)&(df_gps.start_time>=start_time)].next_hop_distance.sum()
        travel_time=patch_distance/speed

        #added time and others
        time_delta=round(start_estimated_stay_duration+travel_time)
        pred_next_start_time=get_time_plus_delta(start_time,time_delta)
        estimated_start_time.append(pred_next_start_time)

        #error cal in Minutes
        error_in_min=(pd.to_datetime(pred_next_start_time).timestamp()-pd.to_datetime(next_start_time).timestamp())/60
        error_in_mins.append(error_in_min)

        #past stay ground_truth
        total_stay+=start_stay_duration

        #speed update
        prev_time=get_time_plus_delta(next_start_time,(-cal_speed_on_past_min*60)) #past 20 min speed is calculated
        df_past_patch=df_gps[(df_gps.start_time<=next_start_time)&(df_gps.start_time>=prev_time)].copy()

        speed=df_past_patch.next_hop_distance.sum()/(df_past_patch.time_elapsed.sum()-total_stay)
        start_time,start_lat,start_long,start_estimated_stay_duration,start_stay_duration=\
        next_start_time,next_lat,next_long,next_estimated_stay_duration,next_stay_duration
        SPEED.append(speed)

    df_mu['predicted_start_time']=estimated_start_time
    df_mu['error_in_min']=error_in_mins
    df_mu['speed']=SPEED

    df_processed=pd.concat([df_mu[['lat','long','speed','start_time','predicted_start_time','error_in_min']],mu_labels],axis=1)
    return df_processed

get_immediate_predicted_arrival_time(test_file_pairs[0])

# Get Prev Egde speed (mean) in meter/sec

for i,(gps_file,mu_file) in enumerate(test_file_pairs):

    pred_time_from_prev=get_immediate_predicted_arrival_time((gps_file,mu_file))
    
    tz=pred_time_from_prev.time_zone[0] #Time Zone    
    down_or_up=gps_file.split("\\")[1] #down / up
    date='-'.join(gps_file.split("\\")[2].split("_")[0].split("-")[::-1]) #date
    
    file_to_save_in_struct=f"pred_time_from_prev_{i}_{date}.csv"
    file_to_save_in_folder=f"{down_or_up}_{tz}_pred_time_from_prev_{i}_{date}.csv"
    
    print("Saving",i,"trail")
    pred_time_from_prev.to_csv(gps_file.split("GPS")[0]+file_to_save_in_struct,index=False)
    pred_time_from_prev.to_csv("./report/test_files/"+file_to_save_in_folder,index=False)

import os

for f in glob.glob("./Trails/*/*/pred_time_from_prev*.csv"):
    os.remove(f)

glob.glob("./Trails/*/*/pred_time_from_prev*.csv")

#NICE
